Update README.md
This project is an application of IoT (Internet of Things) by sending temperature and humidity reading data from the DHT22 sensor with NodeMCU to other NodeMCU via the MQTT mosquitto server. In sending data, encryption and decryption will be carried out from AES cryptography. Later it will be analyzed how the characteristics of data transmission through MQTT.

Library used:
For DHT22
https://github.com/adafruit/DHT-sensor-library
For MQTT
https://github.com/knolleary/pubsubclient
For AES
https://github.com/rweather/arduinolibs